pyproject.toml
[project]
name =  "image-processing-toolbox"
authors = [
    {name = "Alexander Carrels"},
]


description = "Sample Python Project for creating a new Python Module"
readme = "README.md"
version = "0.1.0"
requires-python = ">=3.8.1"
dependencies = ["Pillow >= 10.3.0"]


[project.scripts]
resize-images = "resize_image:resize_images"

[project.optional-dependencies]
test = [
    "bandit[toml]==1.7.5", # find common security issues in Python code
    "black==23.3.0",
    "flake8-bugbear==23.5.9",
    "flake8-docstrings",
    "flake8-formatter_junit_xml",
    "flake8",
    "flake8-pyproject",
    "pylint==2.17.4",
    "pylint_junit",
    "pytest-cov==4.0.0",
    "pytest-mock<3.10.1",
    "pytest-runner",
    "pytest==7.3.1",
    "mypy"]

[tool.black]
line-length = 120
fast = true

[tool.coverage.run]
branch = true
omit = [
    "*/__init__.py",
    "src/utils/*",
    ]

[tool.coverage.report]
fail_under = 100

[tool.flake8]
max-line-length = 120